backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0
dj-rest-auth>=5.0.0
celery>=5.3.4
redis>=5.0.1
channels>=4.0.0
channels-redis>=4.1.0
requests>=2.31.0
cryptography>=41.0.0
django-filter>=23.3
Pillow>=10.1.0
whitenoise>=6.6.0
reportlab>=4.0.0
